Add Integrator Notifies

+++ NutriDbService/IntegratorModels/ProfileAddRequest.cs
@@ -5,13 +5,13 @@
         public override string Status => "A3";
         public string ProfileName { get; set; }
         public string Gender { get; set; }
-        public int Age { get; set; }
-        public int WeightKg { get; set; }
-        public int HeightCm { get; set; }
+        public short Age { get; set; }
+        public decimal WeightKg { get; set; }
+        public decimal HeightCm { get; set; }
         public string Goal { get; set; }
-        public int TargetWeightKg { get; set; }
-        public int DailyCaloricNormKcal { get; set; }
+        public decimal TargetWeightKg { get; set; }
+        public decimal DailyCaloricNormKcal { get; set; }
         public string MacronutrientNormG { get; set; }
-        public int WeeklyActivityHours { get; set; }
+        public double WeeklyActivityHours { get; set; }
     }
 }
